+++ apps/backend/src/research/studies/study.entity.ts
@@ -1,4 +1,4 @@
-import { Collection, Entity, OneToMany, Property } from "@mikro-orm/core";
+import { Collection, Entity, Formula, OneToMany, Property } from "@mikro-orm/core";
 import { BaseEntity } from "../../common/entities/base.entity";
 import { Questionnaire } from "../questionnaires/questionnaire.entity";
 
@@ -9,4 +9,7 @@ export class Study extends BaseEntity {
 
   @OneToMany(() => Questionnaire, (questionnaire) => questionnaire.study)
   questionnaires = new Collection<Questionnaire>(this);
+
+  @Formula((alias) => `(select count(*) from questionnaire q where q.study_id = ${alias}.id)`)
+  questionnairesCount? = 0;
 }