+++ apps/frontend/src/routes/_auth/administration/participants/index.tsx
@@ -1,15 +1,21 @@
 import { createFileRoute, Link } from "@tanstack/react-router";
 import { $api } from "../../../../stores/api";
-import { Button, Table } from "@quassel/ui";
+import { Button, Group, Table } from "@quassel/ui";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { $session } from "../../../../stores/session";
 import { useStore } from "@nanostores/react";
 import { format } from "../../../../stores/i18n";
+import { useSort } from "../../../../hooks/useSort";
+import { paths } from "../../../../api.gen";
 
 function AdministrationParticipantsIndex() {
   const { time } = useStore(format);
   const sessionStore = useStore($session);
-  const participants = useSuspenseQuery($api.queryOptions("get", "/participants"));
+
+  const search = Route.useSearch();
+  const { ToggleLink } = useSort(search);
+
+  const participants = useSuspenseQuery($api.queryOptions("get", "/participants", { params: { query: search } }));
   const deleteParticipantMutation = $api.useMutation("delete", "/participants/{id}", {
     onSuccess: () => participants.refetch(),
   });
@@ -18,8 +24,18 @@ function AdministrationParticipantsIndex() {
     <Table>
       <Table.Thead>
         <Table.Tr>
-          <Table.Th>Id</Table.Th>
-          <Table.Th>Birthday</Table.Th>
+          <Table.Th>
+            <Group>
+              Id
+              <ToggleLink sortKey="id" />
+            </Group>
+          </Table.Th>
+          <Table.Th>
+            <Group>
+              Birthday
+              <ToggleLink sortKey="birthday" />
+            </Group>
+          </Table.Th>
           <Table.Th>Actions</Table.Th>
         </Table.Tr>
       </Table.Thead>
@@ -52,6 +68,8 @@ function AdministrationParticipantsIndex() {
   );
 }
 
+type SearchParams = NonNullable<paths["/participants"]["get"]["parameters"]["query"]>;
+
 export const Route = createFileRoute("/_auth/administration/participants/")({
   beforeLoad: () => ({
     actions: [
@@ -65,4 +83,5 @@ export const Route = createFileRoute("/_auth/administration/participants/")({
   }),
   loader: ({ context: { queryClient } }) => queryClient.ensureQueryData($api.queryOptions("get", "/participants")),
   component: () => <AdministrationParticipantsIndex />,
+  validateSearch: (search) => search as SearchParams,
 });